NLvib
added a modified NLvib as an external package.
Now FE models from YetAnotherFEcode can be used together with NLvib.
All the modifications to NLvib are listed in the YetAnotherFEcode_mod_list.txt file. For the time being, the modifications involve only the addition of nonlinearity cases in the HB_residual and shooting_residual functions, plus the addition of the FE_system class.
- added "BeamNLvib" example in the 1D example folder. For now, the code has been tested ONLY for Harmonic Balance and NMA. In the future, this example will be expanded to cover also Frequency Responses and the shooting method.